+++ luda-editor/client/src/pages/sequence_edit_page/components/graphic_list_view/graphic_list_item.rs
@@ -1,0 +1,78 @@
+use crate::{
+    color,
+    pages::sequence_edit_page::components::graphic_list_view::graphic_thumbnail::GraphicThumbnail,
+};
+use namui::prelude::*;
+use namui_prebuilt::{simple_rect, table};
+use rpc::data::ScreenGraphic;
+
+#[component]
+pub struct GraphicListItem<'a> {
+    pub project_id: Uuid,
+    pub wh: Wh<Px>,
+    pub graphic: &'a ScreenGraphic,
+    pub is_selected: bool,
+}
+impl Component for GraphicListItem<'_> {
+    fn render(self, ctx: &RenderCtx) -> RenderDone {
+        const PADDING: Px = px(4.0);
+
+        let Self {
+            project_id,
+            wh,
+            graphic,
+            is_selected,
+        } = self;
+        let graphic_name = match graphic {
+            ScreenGraphic::Image(image) => image.id.to_string(),
+            ScreenGraphic::Cg(cg) => cg.name.clone(),
+        };
+
+        let stroke_color = color::stroke_color(is_selected, false);
+        let stroke_width = match is_selected {
+            true => 2.px(),
+            false => 1.px(),
+        };
+
+        ctx.compose(|ctx| {
+            table::hooks::horizontal([
+                table::hooks::fixed(wh.height, |wh, ctx| {
+                    ctx.add(GraphicThumbnail {
+                        project_id,
+                        wh,
+                        graphic,
+                    });
+                }),
+                table::hooks::ratio(1, |wh, ctx| {
+                    table::hooks::padding(PADDING, |wh, ctx| {
+                        ctx.add(text(TextParam {
+                            text: graphic_name,
+                            x: 0.px(),
+                            y: wh.height / 2.0,
+                            align: TextAlign::Left,
+                            baseline: TextBaseline::Middle,
+                            font: Font {
+                                size: 12.int_px(),
+                                name: "NotoSansKR-Regular".to_string(),
+                            },
+                            style: TextStyle {
+                                color: stroke_color,
+                                ..Default::default()
+                            },
+                            max_width: None,
+                        }));
+                    })(wh, ctx);
+                }),
+            ])(wh, ctx);
+        });
+
+        ctx.component(simple_rect(
+            wh,
+            stroke_color,
+            stroke_width,
+            color::BACKGROUND,
+        ));
+
+        ctx.done()
+    }
+}
